Update2 cargo.hpp

+++ shm/cargo.hpp
@@ -5,16 +5,16 @@
 class Cargo {
   private: 
      std::string name_;
-     size_t amount_;
-     size_t basePrice_;
+     uint8_t amount_;
+     uint8_t basePrice_;
 
   public: 
      Cargo();
-     Cargo(size_t amount, std::string name, size_t basePrice) {}
+     Cargo(uint8_t amount, std::string name, uint8_t basePrice) {}
      ~Cargo() {;} 
 
      std::string getName();
-     size_t getAmount();
-     size_t getBasePrice();
+     uint8_t getAmount();
+     uint8_t getBasePrice();
      bool operator==(Cargo& cargo);
 };
